Reduce translator binary shape logic
